+++ src/main/java/roomescape/service/ReservationWaitingService.java
@@ -10,7 +10,7 @@ import roomescape.domain.reservationtime.ReservationTime;
 import roomescape.domain.theme.Theme;
 import roomescape.dto.reservationWaiting.ReservationWaitingRequest;
 import roomescape.dto.reservationWaiting.ReservationWaitingResponse;
-import roomescape.dto.reservationWaiting.ReservationWaitingSequence;
+import roomescape.repository.ReservationWaitingSequence;
 import roomescape.exception.InvalidInputException;
 import roomescape.exception.ResourceNotFoundException;
 import roomescape.exception.ReservationTimeNotFoundException;
