+++ src/test/java/roomescape/service/ReservationWaitingServiceTest.java
@@ -22,7 +22,7 @@ import roomescape.domain.reservationtime.ReservationTime;
 import roomescape.domain.theme.Theme;
 import roomescape.dto.reservationWaiting.ReservationWaitingRequest;
 import roomescape.dto.reservationWaiting.ReservationWaitingResponse;
-import roomescape.dto.reservationWaiting.ReservationWaitingSequence;
+import roomescape.repository.ReservationWaitingSequence;
 import roomescape.exception.ExpiredDateTimeException;
 import roomescape.exception.InvalidInputException;
 import roomescape.exception.ReservationTimeNotFoundException;
